As long as you can generate the term from some operation in its index, it works.
It breaks on sequences such as fibbonachi which rely on previous terms.

Originally the plan for this app was a rule-based one which would solve the fibonnachi problem, but has turned into a brute-force function finder.

Next step is to allow functions to "look back" and get the value for previous items, in a way which is still valid when "looking back" from the first item (But this causes questions such as: What comes before the beginning of the fib sequence, for example.)